Add files via upload
This is a game that is played through the console.  
You choose a number 0 through 9, the computer will choose a number 0 through 9, and whichever of the two guesses comes closest to another randomly generated number 0 through 9 wins and adds a tally to their score.  The code keeps track of each round number and its results.
You can play a round of the game by running:

console.log(playGame())

Pass your guess in as an argument and hit enter.